refactor: rename Snowflake to SnowflakeUtil
