+++ src/index.js
@@ -25,8 +25,7 @@ module.exports = {
   Intents: require('./util/Intents'),
   Permissions: require('./util/Permissions'),
   Speaking: require('./util/Speaking'),
-  Snowflake: require('./util/Snowflake'),
-  SnowflakeUtil: require('./util/Snowflake'),
+  SnowflakeUtil: require('./util/SnowflakeUtil'),
   Structures: require('./util/Structures'),
   SystemChannelFlags: require('./util/SystemChannelFlags'),
   UserFlags: require('./util/UserFlags'),
